# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FashinHub E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products/{id}/reviews:
    get:
      operationId: getProductReviews
      tags: [reviews]
      summary: Get reviews for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # WISHLIST
...
  /wishlist/{productId}:
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add product to wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Added

    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove product from wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed

  # USER PROFILE
  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    patch:
      operationId: updateUserProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        avatar:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        createdAt:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string

    # Product
...
    Review:
      type: object
      required: [id, userId, userName, rating, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userAvatar:
          type: ["string", "null"]
        rating:
          type: integer
        title:
          type: ["string", "null"]
        body:
          type: ["string", "null"]
        verified:
          type: boolean
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
        title:
          type: string
        body:
          type: string