# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FashinHub E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  userId: number;
  userName: string;
  /** @nullable */
  userAvatar?: string | null;
  rating: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  body?: string | null;
  verified?: boolean;
  createdAt: string;
}